+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.3
+
+- Added support to optional command `GS r` (transmission of status).
+- Added `end job` command.
+- Added `NetworkPrinter.ensureConnected`.
+
 ## 1.0.2
 
 - Added `PrinterCommand.toString`.